backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
openai==1.54.0
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
